+++ NEWS.md
@@ -1,18 +1,23 @@
 # EGM (development version)
 
 * Updates to `{pkgdown}` settings for documentation organization
-* Simplify and update visualizations using `ggm()` object
-* Debug the color issues when adding a color theme to `ggm()`
+* Debugging the color issues when adding a color theme to `ggm()`
 * Creation of subclass of `egm` objects called `ecg` to identify 12-lead ECG subtypes
 
 # EGM 0.1.1
 
 This is an updated release to *CRAN* for this package.
-The minor changes to this version are as follows.
+The changes to this version are as follows.
 
 * Update to `ggm()` function to allow for appropriate theming for dark and light color themes (removes issues with importing colors)
 
 * Update to code and testing suite to assess compatibility with the `{ggplot2}` __v4.0.0__ release
+
+* Addition of a series of functions for F wave extraction and analysis
+
+* Addition of rudimentary functions for windowing signal based on surface ECG
+
+* New function to read in Prucka (CardioLab) for EP study recordings with the `read_prucka()` function
 
 # EGM 0.1.0